Refactor, move responsibility of converting to a buffered image to UnitImageFactory
