+++ game-app/game-core/src/main/java/games/strategy/triplea/image/ImageTransformer.java
@@ -1,6 +1,8 @@
 package games.strategy.triplea.image;
 
 import java.awt.Color;
+import java.awt.Graphics2D;
+import java.awt.Image;
 import java.awt.geom.AffineTransform;
 import java.awt.image.AffineTransformOp;
 import java.awt.image.BufferedImage;
@@ -24,9 +26,21 @@ public class ImageTransformer {
    * Apply color and brightness to the given image. This uses the hue and saturation of the given
    * color. It ignores the brightness of the color and uses the given brightness value.
    */
-  public static void colorize(final Color color, final int brightness, final BufferedImage image) {
+  public static void colorize(final Color color, final int brightness, Image image) {
+    if(!(image instanceof BufferedImage)) {
+      image = convertToBufferedImage(image);
+    }
     final float[] hsb = Color.RGBtoHSB(color.getRed(), color.getGreen(), color.getBlue(), null);
-    colorize((int) (hsb[0] * 360), (int) (hsb[1] * 100), brightness, image);
+    colorize((int) (hsb[0] * 360), (int) (hsb[1] * 100), brightness, (BufferedImage) image);
+  }
+
+  private static BufferedImage convertToBufferedImage(final Image image) {
+    final BufferedImage newImage =
+        new BufferedImage(image.getWidth(null), image.getHeight(null), BufferedImage.TYPE_INT_ARGB);
+    final Graphics2D g = newImage.createGraphics();
+    g.drawImage(image, 0, 0, null);
+    g.dispose();
+    return newImage;
   }
 
   /**
